♻️(courses) make the course run "title" field optional
A course run is perfectly fine without any title so we should allow it
and make sure that course runs display nicely in this case. This will
be useful to automatically synchronize course runs because synchronizing
multilingual titles with an LMS would greatly complicate the feature
with very discutable benefits.